+++ package.json
@@ -21,7 +21,8 @@
     "mongoose": "^6.9.1",
     "morgan": "1.10.0",
     "multer": "^1.4.5-lts.1",
-    "passport-jwt": "^4.0.1"
+    "passport-jwt": "^4.0.1",
+    "uuid": "^9.0.0"
   },
   "devDependencies": {
     "eslint": "7.19.0",
